% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crd_shiny.R
\name{buildShiny.crdPlan}
\alias{buildShiny.crdPlan}
\title{Dashboard App for the completely randomized design (CRD)}
\usage{
\method{buildShiny}{crdPlan}(obj)
}
\arguments{
\item{obj}{an object inheriting from class "crdPlan" used to build
the Shiny app.}
}
\description{
Constructs and starts a Shiny application from an object
  of class "crdPlan".
}
\examples{
\dontrun{
treat <- LETTERS[seq( from = 1, to = 10 )]
rep <- rep(6, 10)
d1 <- crdPlan(treat, rep)
buildShiny(d1)
}

}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@usp.br}
}
