% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbd_shiny.R
\name{buildShiny.rcbdPlan}
\alias{buildShiny.rcbdPlan}
\title{Dashboard App for the Randomized Complete Block Design (RCBD)}
\usage{
\method{buildShiny}{rcbdPlan}(obj)
}
\arguments{
\item{obj}{an object inheriting from class "rcbdPlan" used to build
the Shiny app.}
}
\description{
Constructs and starts a Shiny application from an object
  of class "rcbdPlan".
}
\examples{
\dontrun{
blocks <- paste("Block", seq(1, 6, 1))
treat <- LETTERS[seq( from = 1, to = 7)]
design2 <- rcbdPlan(treat = treat, blocks = blocks)
buildShiny(design2)
}

}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@usp.br}
}
