% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildShiny.R
\name{buildShiny}
\alias{buildShiny}
\title{Dashboard App for the Randomized Complete Block Design (RCBD)}
\usage{
buildShiny(obj)
}
\arguments{
\item{obj}{an object inheriting from class \code{planExp} used to build
the Shiny app.}
}
\description{
Constructs and starts a Shiny application from an object
  of class \code{planExp}.
}
\examples{
\dontrun{
treatMeans <- c(50, 60, 61, 14, 5, 68, 52)
blocks <- paste("Block BB", seq(1, 6, 1))
treat <- LETTERS[seq( from = 1, to = 7)]
design2 <- rcbdPlan(treat = treat, blocks = blocks,
                    treatMeans = treatMeans)
buildShiny(design2)
}
}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@usp.br}
}
