% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crdPlan.R
\name{crdPlan}
\alias{crdPlan}
\title{Completely randomized design (CRD)}
\usage{
crdPlan(treat, rep, seed)
}
\arguments{
\item{treat}{\code{\link[base]{numeric}} or complex vector
containing treatments levels.}

\item{rep}{\code{\link[base]{numeric}} or complex vector containing
thenumber of
replications by treatment.}

\item{seed}{A single \code{\link[base]{numeric}} value, interpreted
as an integer, that specifies the starting value of the random
number generator.}
}
\description{
levels of treatment are randomly assigned to the
  experimental units. Through randomization, every experimental unit
  has same probability of receiving any treatment. This
  provides a basis for making a valid estimate of random
  error which is so essential in testing of significance of
  genuine differences
}
\examples{
## 3 treatments with same number of replicates
crdPlan(treat = 3, rep = 4)

## 4 treatments with different number of replicates
crdPlan(treat = 4, rep = c(rep(4, 3), 6))

## Running the shiny app
treat <- LETTERS[seq( from = 1, to = 10 )]
design <- crdPlan(treat = treat, rep = 6)
\dontrun{
buildShiny(design)
}

## Treatment with different replicates
design2 <- crdPlan(treat = paste("T", seq(1, 7, 1)),
                   rep = c(rep(6, 4),rep(7, 3)))
\dontrun{
buildShiny(design2)
}

}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@usp.br}
}
