% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{gen.list}
\alias{gen.list}
\title{Internal function to build the data}
\description{
This is an internally called function used to build
  data. This function was adapted from \code{fac.gen} function of
  package dae.
}
\keyword{internal}
