% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{generator}
\alias{generator}
\title{Internal function to build the data}
\description{
This is an internally called function used to build
  data adapted from package from \code{fac.gen} function of
  package dae.
}
\keyword{internal}
