% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crdPlan.R
\name{print.crdPlan}
\alias{print.crdPlan}
\title{Print an \code{crdPlan} object}
\usage{
\method{print}{crdPlan}(x, digits, ...)
}
\arguments{
\item{x}{an object inheriting from class
\code{\link[planExp]{crdPlan}}, representing a randomization for
Completely randomized design.}

\item{digits}{a non-null value for \code{digits} specifies the
minimum number of significant digits to be printed in
values. Default \code{NULL}.}

\item{...}{further arguments passed to \code{{\link{print}}}.}
}
\value{
an object inheriting from class \code{print.crdPlan}.
}
\description{
Prints information about randomization of Completely
  randomized design.
}
\examples{
treat <- LETTERS[seq( from = 1, to = 10 )]
rep <- rep(6, 10)
print(crdPlan(treat = treat, rep=rep))
}
\seealso{
\code{\link[planExp]{crdPlan}},
  \code{\link[planExp]{buildShiny}}
}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@usp.br}
}
