% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbdPlan.R
\name{print.rcbdPlan}
\alias{print.rcbdPlan}
\title{Print an \code{rcbdPlan} object}
\usage{
\method{print}{rcbdPlan}(x, digits, ...)
}
\arguments{
\item{x}{an object inheriting from class
\code{\link[planExp]{rcbdPlan}}, representing a randomization for
Completely randomized design.}

\item{digits}{a non-null value for \code{digits} specifies the
minimum number of significant digits to be printed in
values. Default \code{NULL}.}

\item{...}{further arguments passed to \code{{\link{print}}}.}
}
\value{
an object inheriting from class \code{print.rcbdPlan}.
}
\description{
Prints information about randomization of Completely
  randomized design.
}
\examples{
treat <- paste("T", seq(1, 5, 1))
blocks <- 6
print(rcbdPlan(treat, blocks))

}
\seealso{
\code{\link[planExp]{rcbdPlan}},
  \code{\link[planExp]{buildShiny}}
}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@usp.br}
}
