% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbdCheck.R
\name{rcbdCheck}
\alias{rcbdCheck}
\title{Internal Function to check the arguments}
\description{
This is an internally called function to get an initial
  check
}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@usp.br}
}
\keyword{internal}
