% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbdPlan.R
\name{rcbdPlan}
\alias{rcbdPlan}
\title{Randomized Complete Block Design (RBD)}
\usage{
rcbdPlan(treat, blocks, seed)
}
\arguments{
\item{treat}{\code{\link[base]{numeric}} or complex vector
containing treatments levels.}

\item{blocks}{\code{\link[base]{numeric}} or complex vector
containing blocks levels.}

\item{seed}{A single \code{\link[base]{numeric}} value, interpreted
as an integer, that specifies the starting value of the random
number generator.}
}
\description{
levels of treatment are randomly assigned to the
  experimental units within blocks. We assuming that blocks have a
  systematic effect on the statistical comparisons among
  treatements. Through randomization, every experimental unit within
  block has same probability of receiving any treatement. The word
  'complete' indicates that each block (group) contains all
  treatments.
}
\examples{
## 3 treatments and 4 blocks
rcbdPlan(treat = 3, blocks = 4)

## Running the shiny app
treat <- LETTERS[seq( from = 1, to = 10 )]
design <- rcbdPlan(treat = treat, block = 6)
\dontrun{
buildShiny(design)
}

## Priori for treatment means
blocks <- paste("Block ", seq(1, 6, 1))
treat <- LETTERS[seq( from = 1, to = 6 )]
design2 <- rcbdPlan(treat = treat, blocks = blocks)
\dontrun{
buildShiny(design2)
}

}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@usp.br}
}
