% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbdPlan_builder.R
\name{rcbdPlanBuilder}
\alias{rcbdPlanBuilder}
\title{Internal Function to build the dashborad app}
\description{
This is an internally called function used to build the
  dashboard app
}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@usp.br}
}
\keyword{internal}
